Re-Send to same channel support
Users can now send the same (or different) embeds to the same channel over and over. Previously, this worked on mobile only. It seems Discord on Windows doesn't refresh the message cache as often as mobile. This change forces it to refresh the string select menu by rebuilding the message without it, waiting 500ms, and rebuilding it again with a fresh menu.